refactor: Move AccountSelector from Settings into a component
